LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-subclasses/cssPositionValue-expected.txt

CONSOLE MESSAGE: TypeError: CSS.px is not a function. (In 'CSS.px(1)', 'CSS.px' is undefined)

Harness Error (FAIL), message = TypeError: CSS.px is not a function. (In 'CSS.px(1)', 'CSS.px' is undefined)
